Use Module#ancestors, *_instance_methods instead of extensions.
We lose the ability to see origin classes as distinct from their
original classes, but ruby 2.3 no lnoger lets us do so. This is not a
great loss as the original classes should always be empty anyway.

We still need internal_undefined_instance_methods as there is no ruby
counterpart.